Hit moves to .gbptree
